+++ service/conversations/v1/services/api_op_create.go
@@ -1,0 +1,52 @@
+// Package services contains auto-generated files. DO NOT MODIFY
+package services
+
+import (
+	"context"
+	"net/http"
+	"time"
+
+	"github.com/RJPearson94/twilio-sdk-go/client"
+)
+
+// CreateServiceInput defines the input fields for creating a new service resource
+type CreateServiceInput struct {
+	FriendlyName string `validate:"required" form:"FriendlyName"`
+}
+
+// CreateServiceResponse defines the response fields for the created service
+type CreateServiceResponse struct {
+	AccountSid   *string    `json:"account_sid,omitempty"`
+	DateCreated  time.Time  `json:"date_created"`
+	DateUpdated  *time.Time `json:"date_updated,omitempty"`
+	FriendlyName string     `json:"friendly_name"`
+	Sid          string     `json:"sid"`
+	URL          string     `json:"url"`
+}
+
+// Create creates a new service
+// See https://www.twilio.com/docs/conversations/api/service-resource#create-a-service-resource for more details
+// Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+func (c Client) Create(input *CreateServiceInput) (*CreateServiceResponse, error) {
+	return c.CreateWithContext(context.Background(), input)
+}
+
+// CreateWithContext creates a new service
+// See https://www.twilio.com/docs/conversations/api/service-resource#create-a-service-resource for more details
+func (c Client) CreateWithContext(context context.Context, input *CreateServiceInput) (*CreateServiceResponse, error) {
+	op := client.Operation{
+		Method:      http.MethodPost,
+		URI:         "/Services",
+		ContentType: client.URLEncoded,
+	}
+
+	if input == nil {
+		input = &CreateServiceInput{}
+	}
+
+	response := &CreateServiceResponse{}
+	if err := c.client.Send(context, op, input, response); err != nil {
+		return nil, err
+	}
+	return response, nil
+}